auto-start-stop/requirements.txt
croniter==0.3.19
python-dateutil==2.6.1